airflow/requirements-airflow.txt
# Apache Airflow 3.0 e dependências
apache-airflow==3.0.0
apache-airflow-providers-google==10.12.0
apache-airflow-providers-postgres==5.7.1

# Dependências do projeto DGU
pandas>=2.0.0
requests>=2.28.0
google-cloud-bigquery>=3.4.0
pandas-gbq>=0.19.0

# dbt e adaptadores
dbt-core>=1.6.0
dbt-bigquery>=1.6.0

# Dependências para web scraping
beautifulsoup4>=4.11.0
lxml>=4.9.0
html5lib>=1.1

# Utilitários
python-dateutil>=2.8.0
psycopg2-binary>=2.9.0

# Dependências do sistema
setuptools>=65.0.0
wheel>=0.38.0
